+++ tests/helpers/start-app.js
@@ -1,6 +1,8 @@
 import Ember from 'ember';
 import Application from '../../app';
 import config from '../../config/environment';
+import goNewFormHelper from 'ember-flexberry/test-helpers/go-to-new-form';
+import checkOlvConfigHelper from 'ember-flexberry/test-helpers/check-olv-config';
 
 export default function startApp(attrs) {
   let application;
@@ -11,8 +13,14 @@ export default function startApp(attrs) {
   Ember.run(() => {
     application = Application.create(attributes);
     application.setupForTesting();
+    initTestHelpers();
     application.injectTestHelpers();
   });
 
   return application;
+}
+
+function initTestHelpers() {
+  goNewFormHelper();
+  checkOlvConfigHelper();
 }