+++ addon/components/flexberry-menu.js
@@ -29,6 +29,17 @@ import FlexberryBaseComponent from './flexberry-base-component';
         icon: 'trash icon',
         title: 'Delete',
       }],
+    }, {
+      title: 'Item',
+      buttons: [{
+        title: 'edit',
+        buttonClasses: 'icon',
+        iconClass: 'edit icon',
+        disabled: false,
+        buttonAction: () => {
+          window.alert('clicked edit button');
+        }
+      }]
     }],
     ...
     actions: {
@@ -216,6 +227,11 @@ export default FlexberryBaseComponent.extend({
   _onClickHandler(e) {
     // Find clicked menu item element.
     let itemElement = Ember.$(e.target);
+
+    if (itemElement.parent().is('button') || itemElement.is('button')) {
+      return;
+    }
+
     if (!itemElement.hasClass('flexberry-menuitem')) {
       itemElement = itemElement.parents('.flexberry-menuitem');
     }