+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* Button to clear the filter in a column in the `flexberry-objectlistview` and `flexberry-simpleolv` components.
 
 ## [2.5.0-beta.8] - 2020-08-07
 ### Changed